requirements.txt
Django==6.0.1
python-dotenv==1.2.1
Unidecode==1.4.0
Pillow==12.1.0
django-imagekit==6.0.0
djangorestframework==3.16.1
djoser==2.3.3
pytest-django==4.11.1
drf-spectacular==0.29.0
flake8==7.3.0
ruff==0.14.14